Don't use extra memory storing images zoomed in over 100% (#7)
This lets me view large files without running out of memory! For example, a C++ file I'm working on these days is 1300 lines long, and would OOM with the old version of this code but no longer does. Similarly, I'm now able to open a 4400-line Go file, and although that does take 1.1 gigabytes of RAM, it nonetheless works when it didn't previously!

This uses 15x less memory to store the images!


* don't store pyramid levels zoomed in more than 100%, needs cleanup

* debugging, not done yet...

* Add in debugging info; can't find the bug yet...

* it works, need to remove debugging info

* remove print statements

* use bit shifts, rather than multiplying by a power of 2

* tweak: add comment

* remove commented-out code