Support tab formatting correctly, and other bugfixes when displaying context (#11)
One big part here is that if you change the text width with `-tw`, it changed the width of the window but not the width of the text displayed! :flushed: So, that's now fixed. 

I also remember putting in these list comprehensions where I didn't care about the list itself, but only the side effects performed while constructing it. Why bother with the list at all!? Just make it a for loop.


* change the width of text displayed to match the width of the window

* use a for loop instead of a list comprehension. Why did I do it like that originally?

* trying to get tab alignment working, but so far it's not

* remove unused field

* fix the tabs properly. It works this time

* use the named constant everywhwere

* make the constant a part of the class

* go back to the original phrasing

* tweak comment