Tweaks, bugfixes, and optimizations (#17)
This fixes a subtle problem. Python's garbage collector is based on reference counting, and when a `_SegmentUnionFind` stores a reference to itself in `self._root`, it cannot be garbage collected. So, the changes to that are able to reclaim a whole bunch of memory after we're done colorizing the image.

In addition, this PR adds a useful optimization: segments of size 1 can only merge with their immediate neighbors. So, while initializing, grow the segment as long as possible along a diagonal, and if it's only 1 pixel long, don't bother creating a `_SegmentUnionFind` for it. This significantnly cuts down the amount of memory used.

I also sped up execution time. An example benchmark: when run on `local_robot_test.go` (a 3,400 line file containing 28,000 tokens, resulting in a 780 megapixel image), it takes a little under 3 minutes to initialize everything, and at the most uses 10.2 GB of memory (though after initializing everything, most of that gets freed again). That feels tolerable: it's resource-intensive, but it's a very large file. On more typical files, it takes a few seconds at most, and uses a gigabyte at most.

Finally, I split the initialization out to a separate function, because the function was getting so long.


* remove unused import

* don't store unused-looking variables, but document why they're 'unused'

* help the garbage collector: don't store a reference to self

* use less memory: only make UnionFinds for nontrivial pixels

* reorder if statements so the most common one is at the top

* minor tweaks: don't bother scoring trivial or on-main-diagonal segments

* refactor segment initialization to a separate function

* reflow comments for Python's 80-character width

* specify slots in _SegmentUnionFind. Not a bad idea, but not very useful

* switching from a dense numpy array of data structures to a sparse dict

* print how big things are before computing it all

* simplify interface for _SegmentUnionFind constructor

* start correcting docstrings

* rename segment_matrix -> pixel_to_segment

* redo more docstrings

* initialize final scores to 0 or 1, as relevant

* Initialize scores outside of segments more straightforwardly

* get_lengths should be public, in anticipation of generating automated reports

* iterate over existing segments when setting scores, not places scores might be set

* update docstring