Indent/dedent bugfixes (#19)
Indenting had an off-by-one error when highlighting the token in the context.

Dedenting did not highlight tokens at all, and now does.

Finally, the main file is starting to get long and do multiple things, so I moved all the tokenizing stuff to a new file, and then refactored it into smaller pieces.


* fix off-by-one error with python indent highlighting

* highlight dedent tokens correctly in Python

* debugging

* move make_matrix() into utils.py

* move get_tokens to a separate file

* move FileInfo into tokenizer.py

* remove unused import and unnecessary argument

* whitespace tweak

* split get_boundaries() to a standalone function

* simplify code when parsing dedents

* fix off-by-one error

* more refactoring and fiddling

* rename variable t -> tok

* pull _find_boundary out to a standalone function

* update docstring

* tweak

* redo docstring again

* docstring tweak

* use a match statement instead of an elif chain, not sure if good idea

* tweak