Add integration tests (#22)
The plan in here is to get ready to generate reports about entire repos, though I haven't built that yet. but here are the example files I've been working off of so far, and some tests to make sure I haven't broken image-generation.

I found a surprising bug: when merging segments, I was iterating through a set, and there are times when the order in which you merge them becomes important! So, to prevent the new tests from being flaky, I sorted the segments before merging them.


* attempt to link to the slide I want to reference in Dan Kaminsky's talk

* split out a way to get the segments, separate from the length image

* remove unused variables

* split out code to where it belongs

* rename _get_segments -> _get_pixel_to_segment

* whitespace tweak

* move guess_language to utils.py

* make a helper for getting data from a filename

* add integration test: not done, but already uncovered nondeterminism

* make the code deterministic, so the tests pass

* add a test comparing one file to another

* tweaks

* remember to add the new image, too

* mention in the readme how to run the tests

* remove unused import

* tweak where the comment goes