Prevent flakiness in integration test (#23)
Surprise! My two computers somehow got different tokenizers for Go. Consider this example:
```
`trouble`
```
One would tokenize that as a single token containing everything from one backtick to the other (including both backticks), and the other would tokenize it as 3 separate tokens: a backtick, a string containing `"trouble"`, and another backtick.

Rather than figure out if this is because tree-sitter updated something, or because one of my machines uses go 1.21 and the other uses go 1.20, or it's OS-dependent, or what, the simple thing to do is to not use backticks in the integration test. :upside_down_face: 


* write extra stuff

* redo GPS (go) images without backticks

* Revert "write extra stuff"

This reverts commit 6e6e02f1a5ab889aff8233db53a6404582b24dcb.